requirements.txt
langchain==0.0.154
PyPDF2==3.0.1
python-dotenv==1.0.0
streamlit==1.18.1
faiss-cpu==1.7.4
altair<5
openai==0.28
tiktoken
streamlit-on-Hover-tabs
streamlit-extras
streamlit-chat 
duckduckgo-search
docarray
pypdf